+++ packages/g-webgl/src/utils/dom.ts
@@ -2,7 +2,8 @@ import { isBrowser } from './browser';
 
 const PX_SUFFIX = 'px';
 export function setDOMSize($el: HTMLElement | OffscreenCanvas, width: number, height: number) {
-  if (isBrowser) {
+  // @ts-ignore
+  if (isBrowser && $el.style) {
     // @ts-ignore
     $el.style.width = width + PX_SUFFIX;
     // @ts-ignore